requirements.txt
psycopg[binary]
asyncio
aiopg